edge/requirements.txt
opencv-python>=4.9.0.80
ultralytics>=8.2.0
numpy>=1.26.4
requests>=2.32.3
pyyaml>=6.0.1
